fix encoder definition bug

+++ Arduino_code/encoder_funcs.hpp
@@ -8,9 +8,9 @@ namespace enc_f
     // Wheel radius (in mm)
     extern const int WHEEL_RADIUS;
 
-    long get_ticks(Encoder& e,float d);
-    int get_speed_inch(Encoder& e, float d);
-    int get_speed_mm(Encoder& e, float d);
+    long get_ticks(Encoder& e,int d);
+    int get_speed_inch(Encoder& e, int d);
+    int get_speed_mm(Encoder& e, int d);
 } // enc_f namespace
 
 #endif //ENC_FUNC
